Home work: inheritance - update 5
