--- src/main/java/atUniProMaven/inheritance/Dictionary.java
@@ -1,22 +1,0 @@
-package atUniProMaven.inheritance;
-
-public class Dictionary {
-
-    public enum MessageType {
-        POST("Post"),
-        REPOST("Repost"),
-        COMMENT("Comment"),
-        DIAGNOSIS_TEXT("Diagnosetext");
-
-        private String name;
-
-        MessageType(String name) {
-            this.name = name;
-        }
-
-        public String getName() {
-            return name;
-        }
-    }
-
-}
